requirements.txt
uuid
python-dotenv
pathlib
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
unstructured
chromadb
huggingface_hub
sentence-transformers
